chore: fixes public api and compat tests
