+++ packages/grafana-plugin-compat/src/apps.ts
@@ -5,16 +5,12 @@ import {
   updateAppPluginSettings as runtimeUpdateAppPluginSettings,
 } from '@grafana/runtime/unstable';
 
-export function getPluginSettings(
-  pluginId: string,
-  showErrorAlert?: boolean,
-  getFn: typeof runtimeGetPluginSettings = runtimeGetPluginSettings
-): Promise<PluginMeta> {
-  if (!getFn || typeof getFn !== 'function') {
-    return backwardsCompatibleGetPluginSettings(pluginId, showErrorAlert);
+export function getPluginSettings(pluginId: string, showErrorAlert = false): Promise<PluginMeta> {
+  if (runtimeGetPluginSettings && typeof runtimeGetPluginSettings === 'function') {
+    return runtimeGetPluginSettings(pluginId, showErrorAlert);
   }
 
-  return runtimeGetPluginSettings(pluginId, showErrorAlert);
+  return backwardsCompatibleGetPluginSettings(pluginId, showErrorAlert);
 }
 
 function backwardsCompatibleGetPluginSettings(pluginId: string, showErrorAlert = false): Promise<PluginMeta> {
@@ -24,16 +20,12 @@ function backwardsCompatibleGetPluginSettings(pluginId: string, showErrorAlert =
   });
 }
 
-export function updatePluginSettings(
-  pluginId: string,
-  data: Partial<PluginMeta>,
-  updateFn: typeof runtimeUpdateAppPluginSettings = runtimeUpdateAppPluginSettings
-): Promise<PluginMeta> {
-  if (!updateFn || typeof updateFn !== 'function') {
-    return backwardsCompatibleUpdatePluginSettings(pluginId, data);
+export function updatePluginSettings(pluginId: string, data: Partial<PluginMeta>): Promise<PluginMeta> {
+  if (runtimeUpdateAppPluginSettings || typeof runtimeUpdateAppPluginSettings === 'function') {
+    return runtimeUpdateAppPluginSettings(pluginId, data);
   }
 
-  return runtimeUpdateAppPluginSettings(pluginId, data);
+  return backwardsCompatibleUpdatePluginSettings(pluginId, data);
 }
 
 async function backwardsCompatibleUpdatePluginSettings(
